+++ Include/eval.h
@@ -10,14 +10,24 @@ extern "C" {
 PyAPI_FUNC(PyObject *) PyEval_EvalCode(PyObject *, PyObject *, PyObject *);
 
 PyAPI_FUNC(PyObject *) PyEval_EvalCodeEx(PyObject *co,
-					PyObject *globals,
-					PyObject *locals,
-					PyObject **args, int argc,
-					PyObject **kwds, int kwdc,
-					PyObject **defs, int defc,
-					PyObject *kwdefs, PyObject *closure);
+                                         PyObject *globals,
+                                         PyObject *locals,
+                                         PyObject **args, int argc,
+                                         PyObject **kwds, int kwdc,
+                                         PyObject **defs, int defc,
+                                         PyObject *kwdefs, PyObject *closure);
 
 #ifndef Py_LIMITED_API
+PyAPI_FUNC(PyObject *) _PyEval_EvalCodeWithName(
+    PyObject *co,
+    PyObject *globals, PyObject *locals,
+    PyObject **args, Py_ssize_t argcount,
+    PyObject **kwnames, PyObject **kwargs,
+    Py_ssize_t kwcount, int kwstep,
+    PyObject **defs, Py_ssize_t defcount,
+    PyObject *kwdefs, PyObject *closure,
+    PyObject *name, PyObject *qualname);
+
 PyAPI_FUNC(PyObject *) _PyEval_CallTracing(PyObject *func, PyObject *args);
 #endif
